Layout adjusts + code improvements.
