+++ MoviesUP/Source/Extensions/UIView.swift
@@ -1,0 +1,23 @@
+//
+//  UIView.swift
+//  MoviesUP
+//
+//  Created by Julio Garavelli on 26/11/19.
+//  Copyright © 2019 Julio Garavelli. All rights reserved.
+//
+
+import UIKit
+
+extension UIView {
+
+    func addGradientBackground(firstColor: UIColor, secondColor: UIColor) {
+        clipsToBounds = true
+        let gradientLayer = CAGradientLayer()
+        gradientLayer.colors = [firstColor.cgColor, secondColor.cgColor]
+        gradientLayer.frame = self.bounds
+        gradientLayer.startPoint = CGPoint(x: 0, y: 0)
+        gradientLayer.endPoint = CGPoint(x: 0, y: 1)
+        print(gradientLayer.frame)
+        self.layer.insertSublayer(gradientLayer, at: 0)
+    }
+}
