+++ MoviesUP/Source/View/Collection Cell/ImageCastCollectionViewCell.swift
@@ -13,5 +13,4 @@ class ImageCastCollectionViewCell: UICollectionViewCell {
     @IBOutlet weak var imageView: UIImageView!
     @IBOutlet weak var nameLabel: UILabel!
     @IBOutlet weak var characterNameLabel: UILabel!
-    
 }
